Improve part_2
